Refine seller transaction listing review fixes
- switch the listing hero overlay to light theme styling, center the title, and simplify the pending state
- extract the local seller transaction status icon into its own Item-scoped component
- collapse the preview stack into a folder-local Preview component and keep the transaction-entry query open for later cache reuse
- rename seller transaction status translation keys to a dedicated Transaction status namespace and fill generated locale values
- refresh the seller README to match the updated local Item structure
