+++ apps/app/src/app/@seller/README.md
@@ -41,4 +41,4 @@ Maps to server API: `/api/seller/*`.
 - Seller transaction action buttons now use the seller transaction mutation SDK surface (`accept`, `dispute`, `reject`, `resolve`) and invalidate `transaction-entry` timeline queries.
 - Seller message detail UI now reads conversation timeline through `@user/transaction-entry`, with `resolved` treated as seller read-only.
 - Seller `TitleContainer` pages that navigate back home now use the shared `@common/nav/BackHomeButton` instead of repeating inline `LinkTo + uiBackButton` wiring.
-- Seller listing-level transaction screen now lives in active scope at `@seller/transaction-listing/*`; its hero banner is composed through a local `ListingTransactionHero/*` suspense folder and seller transaction rows now use the local `ui/Item/*` stack to show status-driven content plus client-side last-activity preview instead of repeated listing data.
+- Seller listing-level transaction screen now lives in active scope at `@seller/transaction-listing/*`; its hero banner is composed through a local `ListingTransactionHero/*` suspense folder and seller transaction rows now use the local `ui/Item/*` stack, including folder-local `Preview/*` and `StatusIcon.tsx`, to show status-driven content plus client-side last-activity preview instead of repeated listing data.
