+++ apps/app/src/app/@seller/transaction-listing/ui/Item/Preview/index.ts
@@ -1,0 +1,1 @@
+export { Preview } from "./Preview";
